requirements.txt
discord.py
apscheduler
matplotlib
pandas
python-dateutil
notion-client
pytz
python-dotenv
